Modificada a camada de serviço para ter uma interface e sua implementação
